NodesAndModes.Hex

NodesAndModes.Hex.basis — Method

basis(N, r, s, t)

Computes orthonormal basis of degree N at points (r,s,t).

NodesAndModes.Hex.equi_nodes — Method

equi_nodes(N)

Compute equispaced nodes of degree N.

NodesAndModes.Hex.grad_vandermonde — Method

grad_vandermonde(N,r)

Computes the generalized Vandermonde derivative matrix V of degree N at points r. Specialized for the 1D case

NodesAndModes.Hex.grad_vandermonde — Method

grad_vandermonde(N,rst...)

Computes the generalized Vandermonde derivative matrix V of degree N at points (r,s,t).

NodesAndModes.Hex.nodes — Method

nodes(N)

Computes optimized interpolation nodes of degree N.

NodesAndModes.Hex.quad_nodes — Method

quad_nodes(N)

Compute quadrature nodes and weights exact for degree 2N+1 polynomials. Uses a tensor product Gauss quadrature rule.

NodesAndModes.Hex.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at point r. Specialized for the 1D case

NodesAndModes.Hex.vandermonde — Method

vandermonde(N,rst...)

Computes the generalized Vandermonde matrix V of degree N at points (r,s,t).