NodesAndModes.Line

NodesAndModes.Line.basis — Method

basis(N,r)

Computes the generalized Vandermonde matrix V of degree N (along with the derivative matrix Vr) at points r.

NodesAndModes.Line.equi_nodes — Method

equi_nodes(N)

Computes equally spaced nodes of degree N.

NodesAndModes.Line.grad_vandermonde — Method

grad_vandermonde(N,r,s)

Computes the generalized Vandermonde matrix V of degree N at points r.

NodesAndModes.Line.nodes — Method

nodes(N)

Computes interpolation nodes of degree N.

NodesAndModes.Line.quad_nodes — Method

quad_nodes(N)

Computes (N+1)-point Gauss quadrature rule (exact for degree 2N+1 polynomials)

NodesAndModes.Line.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at points r.