NodesAndModes.Pyr

NodesAndModes.Pyr.abctorst — Method

abctorst(a,b,c)

Converts from Stroud coordinates (a,b,c) on [-1,1]^3 to reference element coordinates (r,s,t).

NodesAndModes.Pyr.basis — Function

basis(N,r,s,t,tol=1e-12)

Computes orthonormal semi-nodal basis on the biunit pyramid element.

Warning: nodal derivative matrices may contain errors for nodes at t = 1. A way to avoid this is to use weak differentiation matrices computed using quadrature rules with only interior nodes.

NodesAndModes.Pyr.build_warped_nodes — Method

build_warped_nodes(N,elem::Symbol,r1D)

Computes degree N warp-and-blend interpolation nodes for elem = :Tri, :Pyr, or :Tet based on the 1D node set "r1D". Returns a tuple "rst" containing arrays of interpolation points.

NodesAndModes.Pyr.edge_basis — Method

edge_basis(N, vertices, edges, basis, vertex_functions, rst...)

Computes edge basis given vertex functions and 1D basis.

NodesAndModes.Pyr.edge_basis — Method

edge_basis(N, elem::Symbol, rst...)

returns generalized Vandermonde matrix evaluated using an edge basis.

NodesAndModes.Pyr.equi_nodes — Method

equi_nodes(N)

Computes equispaced nodes of degree N.

NodesAndModes.Pyr.get_edge_list — Method

get_edge_list(elem::Symbol)

Returns list of edges for a specific element (elem = :Tri, :Pyr, or :Tet).

NodesAndModes.Pyr.grad_vandermonde — Method

grad_vandermonde(N,r,s)

Computes the generalized Vandermonde matrix V of degree N at points r.

NodesAndModes.Pyr.interp_1D_to_edges — Method

interp_1D_to_edges(r1D,elem::Symbol)

Interpolates points r1D to the edges of an element (elem = :Tri, :Pyr, or :Tet)

NodesAndModes.Pyr.meshgrid — Method

meshgrid(vx) Computes an (x,y)-grid from the vectors (vx,vx). For more information, see the MATLAB documentation.

NodesAndModes.Pyr.meshgrid — Method

meshgrid(vx,vy,vz) Computes an (x,y,z)-grid from the vectors (vx,vy,vz). For more information, see the MATLAB documentation.

NodesAndModes.Pyr.meshgrid — Method

meshgrid(vx,vy) Computes an (x,y)-grid from the vectors (vx,vy). For more information, see the MATLAB documentation.

NodesAndModes.Pyr.nodes — Method

nodes(N)

Computes interpolation nodes of degree N.

NodesAndModes.Pyr.quad_nodes — Method

quad_nodes(N)

Computes quadrature nodes and weights which are exact for degree 2N polynomials.

NodesAndModes.Pyr.stroud_quad_nodes — Method

stroud_quad_nodes(N)

Computes quadrature nodes and weights which are exact for degree N polynomials.

NodesAndModes.Pyr.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at points r.