NodesAndModes.Tet

NodesAndModes.Tet.basis — Method

basis(N, r, s, t)

Computes orthonormal basis of degree N at points (r,s,t)

NodesAndModes.Tet.grad_vandermonde — Method

grad_vandermonde(N,r)

Computes the generalized Vandermonde derivative matrix V of degree N at points r. Specialized for the 1D case

NodesAndModes.Tet.grad_vandermonde — Method

grad_vandermonde(N,rst...)

Computes the generalized Vandermonde derivative matrix V of degree N at points (r,s,t).

NodesAndModes.Tet.nodes — Method

nodes(N)

Computes interpolation nodes of degree N. Edge nodes coincide with (N+1)-point Lobatto points, while face nodes coincide with Tr.nodes(N).

NodesAndModes.Tet.quad_nodes_tet — Method

quad_nodes_tet(N)

Returns quadrature nodes and weights which exactly integrate degree N polynomials

NodesAndModes.Tet.simplex_3D — Method

simplex_3D(a, b, c, i, j, k)

Evaluate 3D "Legendre" basis phi_ijk at (a,b,c) coordinates on the [-1,1] cube

NodesAndModes.Tet.stroud_quad_nodes — Method

stroud_quad_nodes(N)

Returns Stroud-type quadrature nodes constructed from the tensor product of (N+1)-point Gauss-Jacobi rules. Exact for degree 2N polynomials

NodesAndModes.Tet.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at point r. Specialized for the 1D case

NodesAndModes.Tet.vandermonde — Method

vandermonde(N,rst...)

Computes the generalized Vandermonde matrix V of degree N at points (r,s,t).