NodesAndModes.Wedge

NodesAndModes.Wedge.basis — Function

basis(N, r, s, t)

Computes orthonormal basis of degree N at points (r,s,t)

NodesAndModes.Wedge.equi_nodes — Method

equi_nodes(N)

Compute equispaced nodes of degree N.

NodesAndModes.Wedge.grad_vandermonde — Method

grad_vandermonde(N,r,s)

Computes the generalized Vandermonde matrix V of degree N at points r.

NodesAndModes.Wedge.meshgrid — Method

meshgrid(vx) Computes an (x,y)-grid from the vectors (vx,vx). For more information, see the MATLAB documentation.

NodesAndModes.Wedge.meshgrid — Method

meshgrid(vx,vy,vz) Computes an (x,y,z)-grid from the vectors (vx,vy,vz). For more information, see the MATLAB documentation.

NodesAndModes.Wedge.meshgrid — Method

meshgrid(vx,vy) Computes an (x,y)-grid from the vectors (vx,vy). For more information, see the MATLAB documentation.

NodesAndModes.Wedge.nodes — Method

nodes(N)

Computes interpolation nodes of degree N.

NodesAndModes.Wedge.quad_nodes — Method

quad_nodes(N)

Returns quadrature nodes and weights which exactly integrate degree 2N polynomials.

NodesAndModes.Wedge.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at points r.