NodesAndModes

NodesAndModes.abctorst — Method

abctorst(elem::Pyr,a,b,c)

Converts from Stroud coordinates (a,b,c) on [-1,1]^3 to reference element coordinates (r,s,t).

NodesAndModes.basis — Function

basis(elem::Pyr,N,r,s,t,tol=1e-12)

Computes orthonormal semi-nodal basis on the biunit pyramid element.

Warning: nodal derivative matrices may contain errors for nodes at t = 1. A way to avoid this is to use weak differentiation matrices computed using quadrature rules with only interior nodes.

NodesAndModes.basis — Method

basis(elem::AbstractElemShape, N, rst...)

Computes orthonormal basis of degree N at tuple of coordinate arrays (r,s,t).

NodesAndModes.basis — Method

basis(elem::Line,N,r)

Computes the generalized Vandermonde matrix V of degree N (along with the derivative matrix Vr) at points r.

NodesAndModes.build_warped_nodes — Method

build_warped_nodes(elem::AbstractElemShape, N, r1D)

Computes degree N warp-and-blend interpolation nodes for elem = Tri(), Pyr(), or Tet() based on the 1D node set "r1D". Returns a tuple "rst" containing arrays of interpolation points.

NodesAndModes.edge_basis — Method

edge_basis(elem::AbstractElemShape, N, rst...)

Returns the generalized Vandermonde matrix evaluated using an edge basis (e.g., degree N polynomials over an edge, but linearly blended into the interior). The dimension of the resulting space is simply the number of total nodes on edges of a degree N element.

NodesAndModes.edge_basis — Method

edge_basis(N, vertices, edges, basis, vertex_functions, rst...)

Computes edge basis given vertex functions and 1D basis.

NodesAndModes.equi_nodes — Method

equi_nodes(elem::AbstractElemShape, N)

Compute equispaced nodes of degree N.

NodesAndModes.equi_nodes — Method

equi_nodes(elem::Line,N)

Computes equally spaced nodes of degree N.

NodesAndModes.find_face_nodes — Function

function find_face_nodes(elem, r, s, tol=50*eps())
function find_face_nodes(elem, r, s, t, tol=50*eps())

Given volume nodes r, s, t, finds face nodes. Note that this function implicitly defines an ordering on the faces.

NodesAndModes.gauss_lobatto_quad — Method

gauss_lobatto_quad(α, β, N)

Computes Legendre-Gauss-Lobatto quadrature points and weights with Jacobi weights α,β.

NodesAndModes.gauss_quad — Method

gauss_quad(α, β, N)

Compute nodes and weights for Gaussian quadrature with Jacobi weights (α,β).

NodesAndModes.get_edge_list — Method

get_edge_list(elem::AbstractElemShape)

Returns list of edges for a specific element (elem = Tri(), Pyr(), or Tet()).

NodesAndModes.grad_jacobiP — Method

grad_jacobiP(r, α, β, N, tmp_array=())

Evaluate derivative of Jacobi polynomial (α, β) of order N at points r.

NodesAndModes.grad_vandermonde — Method

grad_vandermonde(elem::AbstractElemShape, N, rst...)

Computes the generalized Vandermonde derivative matrix V of degree N at points (r,s,t).

NodesAndModes.interp_1D_to_edges — Method

interp_1D_to_edges(elem::AbstractElemShape, r1D)

Interpolates points r1D to the edges of an element (elem = :Tri, :Pyr, or :Tet)

NodesAndModes.meshgrid — Method

meshgrid(vx) Computes an (x,y)-grid from the vectors (vx,vx). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl

NodesAndModes.meshgrid — Method

meshgrid(vx,vy,vz) Computes an (x,y,z)-grid from the vectors (vx,vy,vz). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl

NodesAndModes.meshgrid — Method

meshgrid(vx,vy) Computes an (x,y)-grid from the vectors (vx,vy). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl

NodesAndModes.nodes — Method

nodes(elem::AbstractElemShape,N)

Computes interpolation nodes of degree N. Edge nodes coincide with (N+1)-point Lobatto points. Default routine for elem = Tet(), Pyr(), Tri().

For Quad(), Hex(), Wedge() elements, nodes(...) returns interpolation points constructed using a tensor product of lower-dimensional nodes.

NodesAndModes.nodes — Method

nodes(elem::Line,N)

Computes interpolation nodes of degree N.

NodesAndModes.nodes — Method

nodes(elem::Pyr,N)

Computes interpolation nodes of degree N. Edge nodes coincide with (N+1)-point Lobatto points. Triangular face nodes coincide with Tri.nodes(N), quadrilateral face nodes coincide with tensor product (N+1)-point Lobatto points.

NodesAndModes.quad_nodes — Method

quad_nodes(elem::AbstractElemShape, N)

Compute quadrature nodes and weights exact for (at least) degree 2N polynomials.

NodesAndModes.quad_nodes — Method

quad_nodes(elem::Line,N)

Computes (N+1)-point Gauss quadrature rule (exact for degree 2N+1 polynomials)

NodesAndModes.quad_nodes_tet — Method

quad_nodes_tet(N)

Returns quadrature nodes and weights which exactly integrate degree N polynomials

NodesAndModes.rstoab — Function

rstoab(r, s, tol = 1e-12)

Converts from reference bi-unit right triangle coordinate (r,s) to polynomial basis evaluation coordinates (a,b) on the domain [-1,1]^2

NodesAndModes.simplex_2D — Method

simplex_2D(a, b, i, j)

Evaluate 2D PKDO basis phi_ij at points (a,b) on the Duffy domain [-1,1]^2

NodesAndModes.simplex_3D — Method

simplex_3D(a, b, c, i, j, k)

Evaluate 3D "Legendre" basis phi_ijk at (a,b,c) coordinates on the [-1,1] cube

NodesAndModes.stroud_quad_nodes — Method

stroud_quad_nodes(elem::AbstractElemShape,N)

Returns Stroud-type quadrature nodes and weights constructed from the tensor product of (N+1)-point Gauss-Jacobi rules. Exact for degree 2N polynomials

NodesAndModes.vandermonde — Method

vandermonde(elem::AbstractElemShape, N, rst...)

Computes the generalized Vandermonde matrix V of degree N at points (r,s,t).