NodesAndModes

NodesAndModes.build_warped_nodes — Method

build_warped_nodes(N,elem::Symbol,r1D)

Computes degree N warp-and-blend interpolation nodes for elem = :Tri, :Pyr, or :Tet based on the 1D node set "r1D". Returns a tuple "rst" containing arrays of interpolation points.

NodesAndModes.edge_basis — Method

edge_basis(N, vertices, edges, basis, vertex_functions, rst...)

Computes edge basis given vertex functions and 1D basis.

NodesAndModes.edge_basis — Method

edge_basis(N, elem::Symbol, rst...)

returns generalized Vandermonde matrix evaluated using an edge basis.

NodesAndModes.gauss_lobatto_quad — Method

gauss_lobatto_quad(α, β, N)

Computes Legendre-Gauss-Lobatto quadrature points and weights.

NodesAndModes.gauss_quad — Method

gauss_quad(α, β, N)

Compute nodes and weights for Gaussian quadrature with Jacobi weights (α,β)

NodesAndModes.get_edge_list — Method

get_edge_list(elem::Symbol)

Returns list of edges for a specific element (elem = :Tri, :Pyr, or :Tet).

NodesAndModes.interp_1D_to_edges — Method

interp_1D_to_edges(r1D,elem::Symbol)

Interpolates points r1D to the edges of an element (elem = :Tri, :Pyr, or :Tet)

NodesAndModes.jacobiP — Method

jacobiP(x, α, β, N)

Evaluate the Jacobi Polynomial (α, β) of order N at points x

NodesAndModes.meshgrid — Method

meshgrid(vx) Computes an (x,y)-grid from the vectors (vx,vx). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl

NodesAndModes.meshgrid — Method

meshgrid(vx,vy,vz) Computes an (x,y,z)-grid from the vectors (vx,vy,vz). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl

NodesAndModes.meshgrid — Method

meshgrid(vx,vy) Computes an (x,y)-grid from the vectors (vx,vy). For more information, see the MATLAB documentation.

Copied and pasted directly from VectorizedRoutines.jl. Using VectorizedRoutines.jl directly causes Pkg versioning issues with SpecialFunctions.jl