NodesAndModes.Pyr

NodesAndModes.Pyr.abctorst — Method

abctorst(a,b,c)

Converts from Stroud coordinates (a,b,c) on [-1,1]^3 to reference element coordinates (r,s,t).

NodesAndModes.Pyr.basis — Function

basis(N,r,s,t,tol=1e-12)

Computes orthonormal semi-nodal basis on the biunit pyramid element.

Warning: nodal derivative matrices may contain errors for nodes at t = 1. A way to avoid this is to use weak differentiation matrices computed using quadrature rules with only interior nodes.

NodesAndModes.Pyr.equi_nodes — Method

equi_nodes(N)

Computes equispaced nodes of degree N.

NodesAndModes.Pyr.grad_vandermonde — Method

grad_vandermonde(N,r,s)

Computes the generalized Vandermonde matrix V of degree N at points r.

NodesAndModes.Pyr.nodes — Method

nodes(N)

Computes interpolation nodes of degree N.

NodesAndModes.Pyr.quad_nodes — Method

quad_nodes(N)

Computes quadrature nodes and weights which are exact for degree 2N polynomials.

NodesAndModes.Pyr.stroud_quad_nodes — Method

stroud_quad_nodes(N)

Computes quadrature nodes and weights which are exact for degree N polynomials.

NodesAndModes.Pyr.vandermonde — Method

vandermonde(N,r)

Computes the generalized Vandermonde matrix V of degree N at points r.